# TruckBooking/src/main/java/com/TruckBooking/biddingApi/Model/BidDeleteResponse.java
package com.TruckBooking.biddingApi.Model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@NoArgsConstructor
@AllArgsConstructor
public @Data class BidDeleteResponse {

	private String status;

}