# TruckBooking/src/main/java/com/TruckBooking/biddingApi/Properties/Swagger.YAML
{
    "swagger": "2.0",
    "info": {
        "description": "Api Documentation",
        "version": "1.0",
        "title": "Api Documentation",
        "termsOfService": "urn:tos",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "tags": [
        {
            "name": "basic-error-controller",
            "description": "Basic Error Controller"
        },
        {
            "name": "bidding-controller",
            "description": "Bidding Controller"
        }
    ],
    "paths": {
        "/bid": {
            "get": {
                "tags": [
                    "bidding-controller"
                ],
                "summary": "getBid",
                "operationId": "getBidUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "loadId",
                        "in": "query",
                        "description": "loadId",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "pageNo",
                        "in": "query",
                        "description": "pageNo",
                        "required": false,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/BiddingData"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "bidding-controller"
                ],
                "summary": "addBid",
                "operationId": "addBidUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "bidPostRequest",
                        "description": "bidPostRequest",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BidPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BidPostResponse"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/bid/{Id}": {
            "get": {
                "tags": [
                    "bidding-controller"
                ],
                "summary": "getBidById",
                "operationId": "getBidByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BiddingData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "bidding-controller"
                ],
                "summary": "deleteBid",
                "operationId": "deleteBidUsingDELETE",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "Id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BidDeleteResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/bid/{id}": {
            "put": {
                "tags": [
                    "bidding-controller"
                ],
                "summary": "updateBid",
                "operationId": "updateBidUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "bidPutRequest",
                        "description": "bidPutRequest",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/BidPutRequest"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BidPutResponse"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/error": {
            "get": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingGET",
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "head": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingHEAD",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingDELETE",
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "options": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingOPTIONS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "patch": {
                "tags": [
                    "basic-error-controller"
                ],
                "summary": "errorHtml",
                "operationId": "errorHtmlUsingPATCH",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ModelAndView"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "definitions": {
        "BidDeleteResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            },
            "title": "BidDeleteResponse"
        },
        "BidPostRequest": {
            "type": "object",
            "properties": {
                "loadId": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer",
                    "format": "int64"
                },
                "transporterId": {
                    "type": "string"
                },
                "truckId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "unitValue": {
                    "type": "string",
                    "enum": [
                        "PER_TON",
                        "PER_TRUCK"
                    ]
                }
            },
            "title": "BidPostRequest"
        },
        "BidPostResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            },
            "title": "BidPostResponse"
        },
        "BidPutRequest": {
            "type": "object",
            "properties": {
                "rate": {
                    "type": "integer",
                    "format": "int64"
                },
                "shipperApproval": {
                    "type": "boolean"
                },
                "transporterApproval": {
                    "type": "boolean"
                },
                "truckId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "unitValue": {
                    "type": "string",
                    "enum": [
                        "PER_TON",
                        "PER_TRUCK"
                    ]
                }
            },
            "title": "BidPutRequest"
        },
        "BidPutResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            },
            "title": "BidPutResponse"
        },
        "BiddingData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "loadId": {
                    "type": "string"
                },
                "rate": {
                    "type": "integer",
                    "format": "int64"
                },
                "shipperApproval": {
                    "type": "boolean"
                },
                "transporterApproval": {
                    "type": "boolean"
                },
                "transporterId": {
                    "type": "string"
                },
                "truckId": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "unitValue": {
                    "type": "string",
                    "enum": [
                        "PER_TON",
                        "PER_TRUCK"
                    ]
                }
            },
            "title": "BiddingData"
        },
        "ModelAndView": {
            "type": "object",
            "properties": {
                "empty": {
                    "type": "boolean"
                },
                "model": {
                    "type": "object"
                },
                "modelMap": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "object"
                    }
                },
                "reference": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ACCEPTED",
                        "ALREADY_REPORTED",
                        "BAD_GATEWAY",
                        "BAD_REQUEST",
                        "BANDWIDTH_LIMIT_EXCEEDED",
                        "CHECKPOINT",
                        "CONFLICT",
                        "CONTINUE",
                        "CREATED",
                        "DESTINATION_LOCKED",
                        "EXPECTATION_FAILED",
                        "FAILED_DEPENDENCY",
                        "FORBIDDEN",
                        "FOUND",
                        "GATEWAY_TIMEOUT",
                        "GONE",
                        "HTTP_VERSION_NOT_SUPPORTED",
                        "IM_USED",
                        "INSUFFICIENT_SPACE_ON_RESOURCE",
                        "INSUFFICIENT_STORAGE",
                        "INTERNAL_SERVER_ERROR",
                        "I_AM_A_TEAPOT",
                        "LENGTH_REQUIRED",
                        "LOCKED",
                        "LOOP_DETECTED",
                        "METHOD_FAILURE",
                        "METHOD_NOT_ALLOWED",
                        "MOVED_PERMANENTLY",
                        "MOVED_TEMPORARILY",
                        "MULTIPLE_CHOICES",
                        "MULTI_STATUS",
                        "NETWORK_AUTHENTICATION_REQUIRED",
                        "NON_AUTHORITATIVE_INFORMATION",
                        "NOT_ACCEPTABLE",
                        "NOT_EXTENDED",
                        "NOT_FOUND",
                        "NOT_IMPLEMENTED",
                        "NOT_MODIFIED",
                        "NO_CONTENT",
                        "OK",
                        "PARTIAL_CONTENT",
                        "PAYLOAD_TOO_LARGE",
                        "PAYMENT_REQUIRED",
                        "PERMANENT_REDIRECT",
                        "PRECONDITION_FAILED",
                        "PRECONDITION_REQUIRED",
                        "PROCESSING",
                        "PROXY_AUTHENTICATION_REQUIRED",
                        "REQUESTED_RANGE_NOT_SATISFIABLE",
                        "REQUEST_ENTITY_TOO_LARGE",
                        "REQUEST_HEADER_FIELDS_TOO_LARGE",
                        "REQUEST_TIMEOUT",
                        "REQUEST_URI_TOO_LONG",
                        "RESET_CONTENT",
                        "SEE_OTHER",
                        "SERVICE_UNAVAILABLE",
                        "SWITCHING_PROTOCOLS",
                        "TEMPORARY_REDIRECT",
                        "TOO_EARLY",
                        "TOO_MANY_REQUESTS",
                        "UNAUTHORIZED",
                        "UNAVAILABLE_FOR_LEGAL_REASONS",
                        "UNPROCESSABLE_ENTITY",
                        "UNSUPPORTED_MEDIA_TYPE",
                        "UPGRADE_REQUIRED",
                        "URI_TOO_LONG",
                        "USE_PROXY",
                        "VARIANT_ALSO_NEGOTIATES"
                    ]
                },
                "view": {
                    "$ref": "#/definitions/View"
                },
                "viewName": {
                    "type": "string"
                }
            },
            "title": "ModelAndView"
        },
        "View": {
            "type": "object",
            "properties": {
                "contentType": {
                    "type": "string"
                }
            },
            "title": "View"
        }
    }
}